+++ products/permissions.py
@@ -8,4 +8,4 @@ class IsAuthorOrReadOnly(permissions.BasePermission):
         if request.method in permissions.SAFE_METHODS:
             return True
         # Write permissions are only allowed to the author of a post
-        return obj.author == request.user
+        return obj.creator == request.user